Add files via upload
I wrote this script back in February. The goal is to have it display the text "LOADING!!!" when you're in a loading screen, and "Not Loading" at all other times. While this script is not very useful on its own, it lays the groundwork for a potential automatic loadless timer through a program like Livesplit.